refactor(geo_nurbs): Unify Foundation Pattern structure with geo_primitives
- Move Extension implementations to *_foundation.rs files
- Align file structure: {shape}.rs (Core + tests) + {shape}_foundation.rs (Extension + tests)
- curve_2d: Move ExtensionFoundation/Bounded from curve_2d.rs to curve_2d_foundation.rs
- surface_3d: Move ExtensionFoundation/Bounded from surface_3d.rs to surface_3d_foundation.rs
- Fix Clippy errors (items_after_test_module, similar_names)
- All tests passing (64 tests in geo_nurbs)
