+++ model/geo_nurbs/src/surface_3d_foundation.rs
@@ -1,12 +1,64 @@
-//! Foundation Pattern integration tests for `NurbsSurface3D`
+//! `NurbsSurface3D` Foundation Pattern 統合
+//!
+//! Extension Traits の実装とテスト
+
+use crate::NurbsSurface3D;
+use crate::Scalar;
+use geo_foundation::{Bounded, ExtensionFoundation, NurbsSurface3DMeasure, PrimitiveKind};
+
+// ============================================================================
+// Extension Foundation 実装
+// ============================================================================
+
+impl<T: Scalar> ExtensionFoundation<T> for NurbsSurface3D<T> {
+    fn primitive_kind(&self) -> PrimitiveKind {
+        PrimitiveKind::NurbsSurface3D
+    }
+
+    fn measure(&self) -> Option<T> {
+        Some(<Self as NurbsSurface3DMeasure<T>>::surface_area(self))
+    }
+}
+
+impl<T: Scalar> Bounded<T> for NurbsSurface3D<T> {
+    type Aabb = geo_core::Aabb3D<T>;
+
+    fn aabb(&self) -> Option<Self::Aabb> {
+        // 制御点ベースの境界ボックスを計算
+        let mut min_x = T::from_f64(f64::INFINITY);
+        let mut min_y = T::from_f64(f64::INFINITY);
+        let mut min_z = T::from_f64(f64::INFINITY);
+        let mut max_x = T::from_f64(f64::NEG_INFINITY);
+        let mut max_y = T::from_f64(f64::NEG_INFINITY);
+        let mut max_z = T::from_f64(f64::NEG_INFINITY);
+
+        for u in 0..self.grid_size().0 {
+            for v in 0..self.grid_size().1 {
+                let point = self.control_point(u, v);
+                min_x = min_x.min(point.x());
+                min_y = min_y.min(point.y());
+                min_z = min_z.min(point.z());
+                max_x = max_x.max(point.x());
+                max_y = max_y.max(point.y());
+                max_z = max_z.max(point.z());
+            }
+        }
+
+        Some(geo_core::Aabb3D::new(
+            geo_core::Point3D::new(min_x, min_y, min_z),
+            geo_core::Point3D::new(max_x, max_y, max_z),
+        ))
+    }
+}
+
+// ============================================================================
+// Foundation Pattern 統合テスト
+// ============================================================================
 
 #[cfg(test)]
 mod tests {
-    use crate::NurbsSurface3D;
-    use geo_foundation::{
-        Bounded, ExtensionFoundation, NurbsSurface3DConstructor, NurbsSurface3DMeasure,
-        NurbsSurface3DProperties, PrimitiveKind,
-    };
+    use super::*;
+    use geo_foundation::{NurbsSurface3DConstructor, NurbsSurface3DProperties};
 
     #[test]
     fn test_constructor_new() {
@@ -230,6 +282,7 @@ mod tests {
     }
 
     #[test]
+    #[allow(clippy::similar_names)]
     fn test_measure_tangent_vectors_at() {
         let control_points = vec![
             vec![(0.0, 0.0, 0.0), (0.0, 1.0, 0.0)],
